+++ src/gprmc.c
@@ -32,7 +32,7 @@
 #include <string.h>
 
 bool nmeaGPRMCParse(const char *s, const size_t sz, nmeaGPRMC *pack) {
-  unsigned int fieldCount;
+  size_t fieldCount;
   char timeBuf[16];
   char dateBuf[8];
   bool v23Saved;
@@ -74,7 +74,7 @@ bool nmeaGPRMCParse(const char *s, const size_t sz, nmeaGPRMC *pack) {
   /* see that there are enough tokens */
   if ((fieldCount != 11) //
       && (fieldCount != 12)) {
-    nmeaError(NMEA_PREFIX_GPRMC " parse error: need 11 or 12 tokens, got %d in '%s'", fieldCount, s);
+    nmeaError(NMEA_PREFIX_GPRMC " parse error: need 11 or 12 tokens, got %lu in '%s'", (long unsigned) fieldCount, s);
     goto err;
   }
